requirements.txt
bottle
sqlalchemy
bottle-sqlalchemy
PyYAML
requests
ansible==1.9.1
